10th attempt
this is my nth attempt to upload this project to github, I have tracked the problem to the pennington floor plan texture that I had made for the project, the file was too large for github and thus each push/publish was rejected, though not after trying for 45 minutes to an hour. After deleting the offending file and removing all traces of it, github desktop was still trying to upload it , from where and why I do not know. I have since recloned the repo and started fresh with a copy of the project that has been purged of the too large file and I hope this works.